% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dansample.R
\name{dansample}
\alias{dansample}
\title{My (Dan Groeneveld) Sample Distribution of 10 Equallly Likely Outcomes}
\usage{
dansample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{= number of data points}

\item{iter}{= number of iterations}

\item{time}{= time interval between plotting iterations}
}
\value{
a barplot showing the relative likelihoods of each outcome
}
\description{
My (Dan Groeneveld) Sample Distribution of 10 Equallly Likely Outcomes
}
\examples{
mysample(n=1000, iter=30, time = 1)
}
