% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My bin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{= number of iterations}

\item{n}{= number of trials}

\item{p}{= probability of success}
}
\value{
a table of successes and a barplot
}
\description{
My bin
}
\examples{
mybin=function(iter=10,n=1000, p=0.7)
}
