% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2dg.R
\name{myboot2dg}
\alias{myboot2dg}
\title{My bootstrap 2}
\usage{
myboot2dg(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{= number of iterations}

\item{x}{= data}

\item{fun}{= function to form a confidence interval for}

\item{alpha}{= (1-alpha)100% gives the confidence for the interval}

\item{cx}{= The size of the text displayed on the graph}

\item{...}{= Optional additional text to be displayed above graph}
}
\value{
returns a graph with the confidence interval and has some hidden output, the confidence interval, function used, and data used to display if necessary.
}
\description{
My bootstrap 2
}
\examples{
myboot2dg(sam, iter = 10000, alpha = 0.05)
}
