% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{MyCI}
\usage{
myci(d, n, a)
}
\arguments{
\item{d}{the data}

\item{n}{the number of data points}

\item{a}{alpha}
}
\value{
a confidence interval for the population mean
}
\description{
MyCI
}
\examples{
myci(d=c(1,2,3), n=3, a =0.05) <-95\%CI
}
