% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{mycltdangroeneveld}
\alias{mycltdangroeneveld}
\title{Central limit approximation function for uniform distribution}
\usage{
mycltdangroeneveld(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{The sample size}

\item{iter}{The number of iterations}

\item{a}{The lower limit for the uniform distribution}

\item{b}{The upper limit for the uniform distribution}
}
\value{
Returns a histogram of the distribution of the sum of uniforms
}
\description{
Central limit approximation function for uniform distribution
}
\examples{
w=mycltdangroeneveld(n=10,iter=10000, a = 0, b= 5)
}
