% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikdg.R
\name{mymaxlikdg}
\alias{mymaxlikdg}
\title{My max likelihood}
\usage{
mymaxlikdg(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function applied to data}

\item{x}{the data}

\item{param}{the range of values for the calculated function}

\item{...}{other values such as labels and titles}
}
\value{
a graph with a maximum likelihood for the specified function
}
\description{
My max likelihood
}
\examples{
mymaxlikdg(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)
}
