% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{the value to which you want to calculate the probability from negative infinity}
}
\value{
returns a plot with shaded area and prints out the probability
}
\description{
myncurve
}
