% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Creates a quadratic model equation to be used in plots}
\usage{
myplot(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Creates a quadratic model equation to be used in plots
}
